+++ tests/meson.build
@@ -1352,6 +1352,7 @@ if catch2_dep.found()
       dependency('OpenSSL'),
       dependency('yams_storage_engine'),
       dependency('yams_crypto'),
+      dependency('yams_memory_sync'),
       dependency('yams_core'),
     ],
     cpp_args: test_cpp_noerror_args + ['-DYAMS_TESTING=1'],
@@ -1368,6 +1369,7 @@ if catch2_dep.found()
       dependency('nlohmann_json'),
       dependency('yams_storage_engine'),
       dependency('yams_crypto'),
+      dependency('yams_memory_sync'),
       dependency('yams_core'),
     ],
     cpp_args: test_cpp_noerror_args + ['-DYAMS_TESTING=1'],
@@ -1384,6 +1386,7 @@ if catch2_dep.found()
       dependency('nlohmann_json'),
       dependency('yams_storage_engine'),
       dependency('yams_crypto'),
+      dependency('yams_memory_sync'),
       dependency('yams_core'),
     ],
     cpp_args: test_cpp_noerror_args + ['-DYAMS_TESTING=1'],
@@ -1400,6 +1403,7 @@ if catch2_dep.found()
       dependency('nlohmann_json'),
       dependency('yams_storage_engine'),
       dependency('yams_crypto'),
+      dependency('yams_memory_sync'),
       dependency('yams_core'),
     ],
     cpp_args: test_cpp_noerror_args + ['-DYAMS_TESTING=1'],
@@ -1416,6 +1420,7 @@ if catch2_dep.found()
       dependency('nlohmann_json'),
       dependency('yams_storage_engine'),
       dependency('yams_crypto'),
+      dependency('yams_memory_sync'),
       dependency('yams_core'),
     ],
     cpp_args: test_cpp_noerror_args + ['-DYAMS_TESTING=1'],
